Added GameMachine, Modes, Scoring, GameLoop
GameMachine currently controls the state of the game. Each mode has all the information that mode needs for state. Scoring added so that the player earns points for time survived and coins collected. Added small animations for text above character when collecting coins. Added events to the player that are triggered when they die or collect coins so the game mode only need to add a listener. Also added death check to player if they get off screen