Added levels to game, updated and fixed moving level script, added objects, changed player death
Added 3 levels showing mechanics and requirements for the game.
Changed moving level script to offset the next generated level correctly using the bounds of the level.
Updated logic for death so the only way an obstacle that does damage will kill is if you are on the far left of the screen